requirements.txt
streamlit
numpy
pandas
seaborn
matplotlib
scikit-learn
joblib
google-generativeai
